refactor ar hooks into their own files
minimize passed dependencies to ease testing
