+++ client/allocrunnerv2/alloc_runner_hooks.go
@@ -7,9 +7,7 @@ import (
 
 	log "github.com/hashicorp/go-hclog"
 	multierror "github.com/hashicorp/go-multierror"
-	"github.com/hashicorp/nomad/client/allocdir"
 	"github.com/hashicorp/nomad/client/allocrunnerv2/interfaces"
-	"github.com/hashicorp/nomad/client/allocwatcher"
 )
 
 // initRunnerHooks intializes the runners hooks.
@@ -19,7 +17,7 @@ func (ar *allocRunner) initRunnerHooks() {
 	// Create the alloc directory hook. This is run first to ensure the
 	// directoy path exists for other hooks.
 	ar.runnerHooks = []interfaces.RunnerHook{
-		newAllocDirHook(hookLogger, ar),
+		newAllocDirHook(hookLogger, ar.allocDir),
 		newDiskMigrationHook(hookLogger, ar.prevAllocWatcher, ar.allocDir),
 	}
 }
@@ -135,79 +133,6 @@ func (ar *allocRunner) destroy() error {
 		if ar.logger.IsTrace() {
 			end := time.Now()
 			ar.logger.Trace("finished destroy hooks", "name", name, "end", end, "duration", end.Sub(start))
-		}
-	}
-
-	return nil
-}
-
-// allocDirHook creates and destroys the root directory and shared directories
-// for an allocation.
-type allocDirHook struct {
-	runner *allocRunner
-	logger log.Logger
-}
-
-func newAllocDirHook(logger log.Logger, runner *allocRunner) *allocDirHook {
-	ad := &allocDirHook{
-		runner: runner,
-	}
-	ad.logger = logger.Named(ad.Name())
-	return ad
-}
-
-func (h *allocDirHook) Name() string {
-	return "alloc_dir"
-}
-
-func (h *allocDirHook) Prerun(context.Context) error {
-	return h.runner.allocDir.Build()
-}
-
-func (h *allocDirHook) Destroy() error {
-	return h.runner.allocDir.Destroy()
-}
-
-// diskMigrationHook migrates ephemeral disk volumes. Depends on alloc dir
-// being built but must be run before anything else manipulates the alloc dir.
-type diskMigrationHook struct {
-	allocDir     *allocdir.AllocDir
-	allocWatcher allocwatcher.PrevAllocWatcher
-	logger       log.Logger
-}
-
-func newDiskMigrationHook(logger log.Logger, allocWatcher allocwatcher.PrevAllocWatcher, allocDir *allocdir.AllocDir) *diskMigrationHook {
-	h := &diskMigrationHook{
-		allocDir:     allocDir,
-		allocWatcher: allocWatcher,
-	}
-	h.logger = logger.Named(h.Name())
-	return h
-}
-
-func (h *diskMigrationHook) Name() string {
-	return "migrate_disk"
-}
-
-func (h *diskMigrationHook) Prerun(ctx context.Context) error {
-	// Wait for a previous alloc - if any - to terminate
-	if err := h.allocWatcher.Wait(ctx); err != nil {
-		return err
-	}
-
-	// Wait for data to be migrated from a previous alloc if applicable
-	if err := h.allocWatcher.Migrate(ctx, h.allocDir); err != nil {
-		if err == context.Canceled {
-			return err
-		}
-
-		// Soft-fail on migration errors
-		h.logger.Warn("error migrating data from previous alloc", "error", err)
-
-		// Recreate alloc dir to ensure a clean slate
-		h.allocDir.Destroy()
-		if err := h.allocDir.Build(); err != nil {
-			return fmt.Errorf("failed to clean task directories after failed migration: %v", err)
 		}
 	}
